Fix home page server component metadata
